requirements.txt
mistralai==0.1.8
numpy==1.24.3
plotly==5.9.0
praw==7.7.1
requests_html==0.10.0
seaborn==0.13.2
gensim==4.3.0
matplotlib==3.8.0
nltk==3.8.1
pandas==2.2.2
Pillow==10.3.0
pyLDAvis==3.4.1
seaborn==0.13.2
streamlit==1.33.0
wordcloud==1.9.3